requirements.txt
asgiref==3.2.10
attrs==19.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
charset-normalizer==2.0.0
constantly==15.1.0
cryptography==3.3.2
daphne==2.5.0
Django==3.0
django-braces==1.15.0
django-embed-video==1.4.10
djangorestframework==3.11.0
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
msgpack==1.0.2
pillow==7.2.0
pyasn1==0.4.8
pyasn1_modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
service-identity==18.1.0
setuptools==50.3.2
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.0
typing_extensions==3.7.4.3
tzdata==2020.1
urllib3==1.25.11
zope.interface==5.1.0
